acabados reviews y comments

+++ laravel/app/Http/Controllers/Api/PlaceController.php
@@ -335,27 +335,35 @@ class PlaceController extends Controller
             ]);
             return response()->json([
                 'success' => true,
-                'data'    => "Review created succesfully"
+                'data'    => $review,
             ], 201);
         }
     }
 
 
     /**
-     * Update the specified resource in storage.
+     * Remove the specified resource from storage.
      *
      * @param  int  $id
+     * @param  int  $id2
      * @return \Illuminate\Http\Response
-     */
-    public function unreview($id){
+    */
+    public function unreview($id, $id2){
         $place = Place::find($id);
-        $review = Review::where('user_id',auth()->user()->id)->where('place_id', $place->id )->first();
+        $review = Review::find($id2);
         if ($review){
-            $review->delete();
-            return response()->json([
-                'success' => true,
-                'data'    => "Review deleted successfully"
-            ], 200);
+            if ($place->user_id == auth()->user()->id || auth()->user()->hasRole(['admin']) || $review->user_id == auth()->user()->id ){
+                $review->delete();
+                return response()->json([
+                    'success' => true,
+                    'data'    => "Review deleted successfully"
+                ], 200);
+            }else{
+                return response()->json([
+                    'success'  => false,
+                    'message' => "ERROR, you can not delete a review that is not yours"
+                ], 500);
+            }
         }else{
             return response()->json([
                 'success'  => false,
